Moved the CPU-side hierarchy generation to a separate C++ testing project for now.
Also adjusted my PositionToMortonCode(...) algorithm.  The conversion to
the range [0,1] has been a bit of a thorn in my side ever since I
started this, but I think that I got it now.  At least on the CPU-side
generation of Z-order curve vertices.  I still need to update the
shader.