Learned about padding structures; begin GenerateBoundingVolumes.comp by copying and pasting
It seems that padding structures out to 16-byte alignment is only
necessary when there is an array of the structures AND the structure
defines a vec* or mat* member, like Particle.  Yes, vec2's included.
Despite only being 8bytes, they require the entire structure that they
are part of to be 16-byte aligned.

If the structure is only defined using primitives, however, like
BvhNode, then no padding is necessary.