Created initial collision detection algorithm via BVH traversal, but there's a problem with tree generation.
There is a corner case somewhere that causes a few internal nodes to
develop a circular dependency.  It still works just fine on my test
particles, but doesn't work well when several hundred particles are
active.  The problem is unpredictable, but is likely based around morton
code data, which is dependent upon position, which is randomized, hence
the unpredictability.