I might have it fixed
The tree was constructing itself by analyzing all the leaves, not just
those that corresponded to active particles.  I am currently feeding the
number of active particles into the BVH tree construction via grabbing a
static instance of the persistant atomic counter and relying on it
having just finished counting the number of active particles.  I tried
sticking a numActiveParticles value into the ParticleBuffer, but
rendering and the atomic coutner buffer went "wonky" and I didn't know
how to fix it.