I think I got it.  Problem summary:
Problem (1): Fixed Yesterday.  I had overlooked the race condition that
was introduced by initializing leaf node data at the beginning of
GenerateBinaryRadixTree.comp.  Fixed by making a shader that ran prior
that was dedicated exclusively to that.
Problem (2): Fixed just now.  My "terminate second thread that enters"
check in GenerateBoundingVolumes.comp was in error.  I don't know why,
but checking a node's _threadEntranceCounter immediately after the
atomic add was a problem.  The result of the write may not have been
visible to other threads yet, but the return value of the atomic counter
is immediately visible.  Fixed by checking if the first atomic add
returned a 0 (the counter's default value) and terminating that thread.