Now I fixed it
After a lot of debugging, I found that there was a small hole in the
"number of active particles" counting.  In ParticleUpdate.comp, the "out
of bonds" check was performed after the "is active?" check, so it was
possible for the atomic counter for active particles to be incremented
for a particle that was just made inactive.

And now I can run without the infinite loop problem.  Yay!

But now particles are gradually disappearing.  Need to figure out why.