Tree traversal working, but performance is quite less than hoped for.
The BVH tree traversal is taking significantly longer than I expected.
At ~4500 particles the framerate is down to single digits.  If the
particles are scattered, such as when there is only 1 particle emitted
per emitter per frame, then the tree is wide, overlaps are few, and
frame rate is high.  The problem comes when there are loads of nearby
particles, so there are many overlaps, tree traversal is deep and goes
off on many branches, and frame rate is low.

I may be able to boost performance by making this shader only bother
with the first particle to collide, and then have another shader that
expressly handles possible collisions.  I can also reduce overlap by
shrinking the particle collision radius.