Finally got BVH geometry generated
It took a few hours to sort out why the geometry wasn't being generated
properly.  The end problem was in the use of incorrect indices in the
new GenerateBvhVertices.comp.  I was looking in a lot of places except
there.  After the problem with ParticleUpdate.comp giving an incorrect
reading on the number of active particles, I'm having trouble trusting
any of my shaders right now, but I should have paid more attention to
the new thing that I introduced.