Lots of cleanup.  Have tempory fix to flickering.
Reducing the particle collision radius reduces the number of overlaps
and thus the depth of the traversals.  This improves performance but
also seems to reduce the flickering.  I did notice that the flickering
was always worst in areas of lots of overlap, so perhaps the problem is
related to depth?  I tried increasing the size of the "node index stack"
array in DetectAndResolveCollisions.comp, but that didn't help.

I still have no idea why the flickering is happening.