Renamed project from 2D_GPU_PCollisionWithGeometry -> 2D_GPU_PCollisionWithBvh
I thought that I was going to start doing particle collisions with
geometry after 2D_GPU_PCollisionWithZOrderCurveRadixSorting, but then I
learned that collision detection via proximity on the Z-order curve is
not right and I needed to generate and traverse a BVH.  That's what this
project became.  So I finally renamed the repository and the VS2015
solution and project files to match.