Renamed the sorting and collision shaders (again) to remove the prefix "_#"
The prefix was causing trouble when I wanted to shuffle around some
responsibility or add or remove a shader.  I liked the prefix so that I
could see at a glance which shader came first just by looking at the
list of shaders, but the trouble it caused when adding or removing a
shader wasn't worth it.