Aaand working!
Collision detection is now working properly.  I have tested it so that
the number of potential collisions is greater than the size of the array
in the ParticlePotentialCollisionsBuffer and there are no array
overwrites.  It really starts to bog down when there are a lot of
particles in a small space (lots of tree traversals), but I believe that
this can be solved by simply shrinking the particle collision radius.
Also, particle color is now determined by the number of potential
collisions.  Sweet.