Collision resolution and particle velocity vector geometry generation are building and running, but not working right.
I now need to get particle velocity vector geometry generation running,
and then I can start to deal with why the collision resolution isn't
working.  I modified the end of ResolveCollisions.comp to simply reverse
the velocity in the event of a collision, so the shader is definitely
running, but I don't know why the collision detection seems to do squat.