Fixed collision resolution.  Particles are now bouncing properly.
After a review of the collision handling on the Gamasutra Pool Hall
Elastic Collisions article, I discovered that my calculations of a1 and
a2 (each particles'  velocity vectors along the line of contact) were
incorrectly being created with the dot product of the particles'
velocity vectors and the line of contact.  It was supposed to be with
the normalized line of contact.