Some debugging stuff to hunt down the cause of "nan"
It looks like, occasionally, two particle positions are equivalent.  The
positions are floats and this is very unlikely, but it is apparently
possible.  In that event, the distSqr between the two particles is 0.
Pass a 0 into inversesqrt(...) and it returns "inf".  Multiply "inf" by
the line-of-contact vector and the result is "nan" in each component.
The "nan" propogates through the rest of the calculation until the end,
when a new velocity is assigned.  On the next update shader, that "nan"
velocity is applied to the particle position.  Somehow (I haven't traced
this down) that "nan" quickly propogates to the rest of the particles,
seemingly in the course of 1-2 frames.  It may have something to do with
the Morton Code calculations.  Anyway, I'll just remember to check for
"nan" distSqr and call it good.  This means avoiding some particle
collisions, but that is acceptable for this simulation.